guns can now be picked up
some code changes and editor edits made for weapon and player

+++ FPS Shooter/Assets/Scripts/GunLists.cs
@@ -7,7 +7,7 @@ public class GunLists : ScriptableObject
 {
     public int shootingDamage;
     public float shootingRate;
-    public float shootingDist;
+    public int shootingDist;
 
     public GameObject gunModel;
     public AudioClip gunBlastAudio;
